Created htd object to use for gear, reconfigured files to use the new object
